+++ frontend/src/components/journals/CustomCalendar.tsx
@@ -1,19 +1,14 @@
 import { useEffect, useState } from 'react';
 import Calendar from 'react-calendar';
 import './CustomCalendar.css'; // Custom CSS for transitions
-import { useLocation, useSearchParams } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 import { queryStringKeys } from '@/constants';
 import PrevMonth from '@/assets/icons/btn-prev-month.svg';
 import NextMonth from '@/assets/icons/btn-next-month.svg';
 import { fetchDogMonthStatistic, period } from '@/api/dogs';
 import { formDate, formDay } from '@/utils/date';
+import useCalendar from '@/hooks/useCalendar';
 
-const getStartOfWeek = (date: Date) => {
-    const startDate = new Date(date);
-    const day = startDate.getDay();
-    const diff = startDate.getDate() - day;
-    return new Date(startDate.setDate(diff));
-};
 const formCalendar = (date: Date) => {
     const year = date.getFullYear();
     const month = date.getMonth() + 1;
@@ -23,14 +18,10 @@ const formCalendar = (date: Date) => {
 export type viewMode = 'week' | 'month';
 
 export default function CustomCalendar() {
-    const [searchParams, setSearchParams] = useSearchParams();
     const location = useLocation();
-    const today = new Date();
-    const [date, setDate] = useState<Date>(new Date());
-    const [currentWeek, setCurrentWeek] = useState<Date[]>([]);
-    const [view, setView] = useState<viewMode>('week');
     const [mark, setMark] = useState<any[]>([]);
-
+    const { toggleViewSwitch, handlePrevMonth, handleNextMonth, today, handleClickDay, date, currentWeek, view } =
+        useCalendar();
     const getStatisticData = async (date: string, period: period) => {
         const params = new URLSearchParams(location.search);
         const dogId = params.get(queryStringKeys.DOGID);
@@ -45,49 +36,6 @@ export default function CustomCalendar() {
         setMark(newArray);
     };
 
-    const handlePrevMonth = async () => {
-        const prevMonth = new Date(date.getFullYear(), date.getMonth() - 1, 1);
-        if (today.getFullYear() >= prevMonth.getFullYear() && today.getMonth() >= prevMonth.getMonth()) {
-            await getStatisticData(formDate(prevMonth), 'month');
-        }
-        setDate(prevMonth);
-    };
-
-    const handleNextMonth = async () => {
-        const nextMonth = new Date(date.getFullYear(), date.getMonth() + 1, 1);
-        if (today.getFullYear() >= nextMonth.getFullYear() && today.getMonth() >= nextMonth.getMonth()) {
-            await getStatisticData(formDate(nextMonth), 'month');
-        }
-        setDate(nextMonth);
-    };
-    const toggleSwitch = () => {
-        if (view === 'month') {
-            setView('week');
-        } else {
-            getStatisticData(formDate(today), 'month');
-            setView('month');
-        }
-    };
-    const handleClickDay = (value: Date) => {
-        const date = formDate(value);
-        setDate(value);
-        if (!mark.includes(date)) return;
-
-        const newSearchParams = new URLSearchParams(searchParams);
-        newSearchParams.set(queryStringKeys.DATE, date);
-        setSearchParams(newSearchParams);
-    };
-
-    useEffect(() => {
-        const startOfWeek = getStartOfWeek(date);
-        const week = Array.from({ length: 7 }).map((_, index) => {
-            const day = new Date(startOfWeek);
-            day.setDate(startOfWeek.getDate() + index);
-            return day;
-        });
-        setCurrentWeek(week);
-    }, [date]);
-
     useEffect(() => {
         getStatisticData(formDate(today), 'week');
     }, []);
@@ -95,13 +43,21 @@ export default function CustomCalendar() {
         <div className="w-full flex flex-col px-[30px] pt-4 justify-center items-center bg-white shadow rounded-bl-2xl rounded-br-2xl overflow-hidden">
             {view === 'month' && (
                 <div className="w-full flex gap-2 justify-end mb-6">
-                    <button onClick={handlePrevMonth}>
+                    <button
+                        onClick={() => {
+                            handlePrevMonth(getStatisticData);
+                        }}
+                    >
                         <img src={PrevMonth} alt="이전달" />
                     </button>
                     <span className="text-center text-neutral-800 text-base font-bold leading-normal">
                         {formCalendar(date)}
                     </span>
-                    <button onClick={handleNextMonth}>
+                    <button
+                        onClick={() => {
+                            handleNextMonth(getStatisticData);
+                        }}
+                    >
                         <img src={NextMonth} alt="다음달" />
                     </button>
                 </div>
@@ -113,7 +69,9 @@ export default function CustomCalendar() {
                 activeStartDate={date}
                 onActiveStartDateChange={() => {}}
                 showNeighboringMonth={view === 'week'}
-                onClickDay={handleClickDay}
+                onClickDay={(value: Date) => {
+                    handleClickDay(value, mark);
+                }}
                 tileDisabled={({ date }) =>
                     view === 'month'
                         ? false
@@ -141,7 +99,12 @@ export default function CustomCalendar() {
                 }}
             />
             <div className="w-full h-[30px]  justify-center items-center inline-flex">
-                <button className="w-[30px] h-1 bg-neutral-200 rounded-sm" onClick={toggleSwitch}></button>
+                <button
+                    className="w-[30px] h-1 bg-neutral-200 rounded-sm"
+                    onClick={() => {
+                        toggleViewSwitch(getStatisticData);
+                    }}
+                ></button>
             </div>
         </div>
     );
